Allow followers to be added to character sheets
In this update, the Datamodels for squads, characters, and monsters have
been split up. The membership data for squads has been repurposed for
use with disciples, which in turn can be added in the same way as squad
members. If the target sheet is not a squad, the dropped sheet
saves the uuid of it's new master. Actors can be updated from their
master's character sheet, though vice versa does not happen
automatically.